refactor: consolidate handler route-building into a single routes.go
Seven small files (17–33 lines each) containing closely related
route-building methods were split across the handler package with no
clear benefit:

  cache_middleware.go, options_middleware.go, mocked_routes.go,
  static_routes.go, script_routes.go, rewrited_routes.go, helpers.go

All methods now live in routes.go alongside their natural context.
No logic was changed; also simplified the nil-guard in createRoute
(removed redundant len checks before ranging empty maps).
